Fleshed out the core domain
