+++ go.mod
@@ -3,6 +3,11 @@ module github.com/ess/kit
 go 1.16
 
 require (
+	github.com/DATA-DOG/godog v0.7.10
+	github.com/ess/jamaica v1.0.3
+	github.com/ess/kennel v0.0.4
+	github.com/ess/mockable v0.2.0
+	github.com/ess/testscope v1.0.0
 	github.com/spf13/cobra v1.2.1
 	github.com/spf13/viper v1.9.0
 )
